Concluído o desafio 1

+++ semana7/cadastro-usuario-api-labenusers/labenusers/src/components/UsersList/UsersList.js
@@ -36,16 +36,18 @@ const UlPrincipal = styled.ul`
 
 export default class UsersList extends React.Component {
   deleteUser = async (id) => {
-    try {
-        await axios.delete(`${baseUrl}/${id}`, headers)
-        
-        this.props.getAllUsers();
-        alert("Usuário deletado com sucesso");
-    } catch (error) {
-        if (error.response.data.message === "User not found.") {
-          alert("Usuário não encontrado");
+    if (window.confirm('Deseja realmente deleter esse usuário?')) {
+        try {
+            await axios.delete(`${baseUrl}/${id}`, headers)
+            
+            this.props.getAllUsers();
+            alert("Usuário deletado com sucesso");
+        } catch (error) {
+            if (error.response.data.message === "User not found.") {
+              alert("Usuário não encontrado");
+            }
+            console.log(error.response.data.message);
         }
-        console.log(error.response.data.message);
     }
   };
